netx/dialer: refactor base dialer to be composable
Part of https://github.com/ooni/probe-engine/issues/125
